app/core/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Settings:
    PROJECT_NAME = os.getenv("PROJECT_NAME", "Smart Docs API")
    API_V1_STR = os.getenv("API_V1_STR", "/api/v1")
    OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")

settings = Settings()
